Update user tests
